mtd: fix mtdraw_erase stub signature
Commit 81737c1d436a (mtd: Fix erasing of devices >4GiB) changed the
prototype of the erase function, but fogot to fix up the stub function
used when no MTD write support is built in.

+++ drivers/mtd/mtdraw.c
@@ -281,7 +281,7 @@ static ssize_t mtdraw_write(struct cdev *cdev, const void *buf, size_t count,
 {
 	return 0;
 }
-static ssize_t mtdraw_erase(struct cdev *cdev, size_t count, loff_t offset)
+static ssize_t mtdraw_erase(struct cdev *cdev, loff_t count, loff_t offset)
 {
 	return 0;
 }
